+++ source/renderer/app/components/staking/delegation-center/DelegationCenterHeader.tsx
@@ -133,9 +133,10 @@ class DelegationCenterHeader extends Component<Props, State> {
     const slot = get(networkTip, 'slot', '-');
     const headingFirst = intl.formatMessage(messages.headingRight);
     const headingSecond = intl.formatMessage(messages.headingLeft);
-    const timeUntilFutureEpoch = currentLocale
-      ? humanizeDurationByLocale(this.state.timeUntilFutureEpoch, currentLocale)
-      : null;
+    const timeUntilFutureEpoch = humanizeDurationByLocale(
+      this.state.timeUntilFutureEpoch,
+      currentLocale
+    );
     const description = intl.formatMessage(messages.description, {
       timeUntilFutureEpoch,
     });