+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # Changelog
+## vNext
+
+### Features
+
+### Fixes
+
+### Chores
+
+- Fixed locale in Storybook ([PR 3042](https://github.com/input-output-hk/daedalus/pull/3042))
 
 ## 5.1.0
 
@@ -9,7 +18,6 @@
 - Updated trezor-connect to v9 ([PR 3038](https://github.com/input-output-hk/daedalus/pull/3038))
 - Added analytics data collection ([PR 2927](https://github.com/input-output-hk/daedalus/pull/2927), [PR 2989](https://github.com/input-output-hk/daedalus/pull/2989), [PR 3003](https://github.com/input-output-hk/daedalus/pull/3003), [PR 3028](https://github.com/input-output-hk/daedalus/pull/3028))
 - Updated LedgerJS to 5.1.0 ([PR 3036](https://github.com/input-output-hk/daedalus/pull/3036))
-- Fix locale in Storybook ([PR 3042](https://github.com/input-output-hk/daedalus/pull/3042))
 
 ## 5.0.0